+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changed
 
 - [TD-1618] Cache improvements. Use td-cache instead of td-perms.
+- [TD-1866] Exclude logic deleted data_structures in catalog navigation, catalog table and filters
 
 ## [3.0.0] 2019-06-25